[WIP]: refactored calibration & startHandle
