+++ backend/.gitignore
@@ -40,3 +40,5 @@ out/
 .kotlin
 
 .env
+
+static/
