+++ frontend/src/utils/speed_meter_script.d.ts
