+++ frontend/src/utils/speed_meter_script.js
